op2: Use xxhdpi dalvik heap values
Currently, we have seen some native crashes related to gralloc on startup. It is
because the minimum dalvik VM size is too low and it crashes many apps like
Nova launcher and Alipay when memory pressure is high.

Use xxhdpi config from Lineage to fix this. No crashes after this tune.